pyproject.toml
[project]
name = "rivet-pro"
version = "2.0.0"
description = "Industrial maintenance AI assistant"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pillow>=10.0",
    # LLM Providers
    "groq>=0.4",
    "google-generativeai>=0.3",
    "anthropic>=0.18",
    "openai>=1.12",
    # Async
    "asyncpg>=0.29",
    "aiohttp>=3.9",
    # Observability
    "langsmith>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.2",
]
phoenix = [
    "arize-phoenix>=4.0",
]
telegram = [
    "python-telegram-bot>=21.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
